Factor out loader config generator
